+++ binance/spot/pay.py
@@ -6,8 +6,8 @@ def pay_history(self, **kwargs):
     https://binance-docs.github.io/apidocs/spot/en/#get-pay-trade-history-user_data
 
     Keyword Args:
-      startTimestamp (int, optional)
-      endTimestamp (int, optional)
+      startTime (int, optional)
+      endTime (int, optional)
       limit (int, optional): default 100, max 100
       recvWindow (int, optional): The value cannot be greater than 60000
     """